requirements.txt
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
openpyxl==3.1.2
google-api-python-client==2.108.0
google-auth==2.25.2
APScheduler==3.10.4 